requirements.txt
Flask==3.0.0
flask-cors==4.0.0
yt-dlp>=2023.0.0
dashscope>=1.14.0
python-dotenv>=1.0.0